Background data and basic UI
Includes background data and some basic UI for the game;
Default data for the game is initialize in GameManager.Start(), fine
tuning of game mechanics is in the works.

To implement model objects, create reference for those objects in the
GameManager (script) which is attached to the Game Controller (object).
To access datas of water sources for implementing models in the
GameManager script, call data.WaterSources[0] (lake),
data.WaterSources[1] (aquifer). Currently, the only available data for
watersources is reserve, and availability, which have getter functions.