Background data and basic UI
Include Background data scripts and basic UI for the game; Default game
variables for the game is initialize in GameManager.Start(), fine tuning
of game mechanics is in the works.
To implement model objects, create reference for those objects in
GameManager(script) which is attached to the Game Controller (object).
To access datas of water sources for implementing models in the
GameManager script, call data.WaterSources[0] (lake),
data.WaterSources[1] (Aquifer).
Currently, the only available data for watersources is reaserve and
availability, which have getter functions.